python/sagemaker/hugginface/background-removal/requirements.txt
Flask
transformers
torch
torchvision
Pillow